Change script to extract boost from subversion.
Use lipo out of the arm tools directory so it knows about armv7s

Update the extract-combine-repack sequence to handle the armv7s architecture.

Update to work with IOS 6.0 SDK.

Note: The repository used to contain a tarball with boost.
It has been rebased and filter-branched out to make the repository smaller.